requirements.txt
streamlit
langchain
langchain-openai
langchain-community
langchain-huggingface
beautifulsoup4
python-dotenv
faiss-cpu
sentence-transformers